requirements.txt
absl-py==1.4.0
alembic==1.11.1
astunparse==1.6.3
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
chardet==3.0.4
charset-normalizer==3.1.0
clean-text==0.6.0
click==8.1.3
colorama==0.4.6
docopt==0.6.2
emoji==1.7.0
filelock==3.12.0
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
flatbuffers==23.5.26
fsspec==2023.5.0
ftfy==6.1.1
gast==0.4.0
google==3.0.0
google-api-core==2.11.0
google-auth==2.18.1
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-language==2.9.1
google-cloud-storage==2.9.0
google-cloud-videointelligence==2.11.1
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
googletrans==3.1.0a0
greenlet==2.0.2
grpcio==1.54.2
grpcio-status==1.54.2
h11==0.9.0
h2==3.2.0
h5py==3.8.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.14.1
hyperframe==5.2.0
idna==3.4
install==1.3.5
itsdangerous==2.1.2
jax==0.4.10
Jinja2==3.1.2
joblib==1.2.0
keras==2.12.0
Keras-Preprocessing==1.1.2
libclang==16.0.0
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.1.0
mpmath==1.3.0
mysqlclient==2.1.1
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pipreqs==0.4.13
proto-plus==1.22.2
protobuf==4.23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post5
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
subject-classification-spanish==0.0.10
sympy==1.12
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformers==4.29.2
typing_extensions==4.6.1
urllib3==2.0.2
wcwidth==0.2.6
Werkzeug==2.3.4
wrapt==1.14.1
yarg==0.1.9
